Add top/bottom padding to make room for Android copy&paste tooltip
Chrome Android shows a copy/paste/... (but not formatting) tooltip above
selected text. For the first line of text, this blocks our formatting
bar, so we add some padding-top to the document to make room for the
tooltip.

Firefox Android also tries to put the tooltip above the selected text,
but messes up (maybe related to the iframe) and puts it around 100px
above it. However, when that is out of bounds for Firefox's window, it
does end up in the correct spot, which happened before this patch for
the first line (covering the formatting bar again). After this patch,
the tooltip appears around 100px above it.